+++ app/models/surveys/survey_question_order.rb
@@ -1,4 +1,4 @@
 class SurveyQuestionOrder < ActiveRecord::Base
-  belongs_to :question, -> { current }
   belongs_to :survey, -> { current }
+  belongs_to :question, -> { current }
 end